==> CODE/20 Model Comparison.Rmd <==
---
title: "Model Comparison"
author: "Hadasch, Schmidt, Piepho"
date: "05 Februar 2018"
output:
  html_document: default
---



```{r setup, include=FALSE}
require(asreml)
options(scipen=4)
#################
dataset <- "aman"
dataset <- "boro"
#################
dat3 <- read.delim(paste(dataset,"_means.txt",sep=""))
dat3 <- dat3[order(dat3$G,dat3$Env),]
dat3$Y <- as.factor(paste(dat3$Y))
```

# Contents of this file:
#### Dataset = `r dataset`
* Models:
    + `G:Env` Standard Model  
    + `at(G):Env` Shukla Model
    + `fa(G):Env` FA Model
    + `diag(fiveyr):G:Env` 5yr diag Model
* Model Outputs (further formatting not shown):
    + AIC: `-2*MODELOBJECT$loglik + 2*length(MODELOBJECT$gammas)`
    + Wald test: `wald(MODELOBJECT, denDF = c("none"),ssType = c("incremental","conditional"))`
    + Variance Components: `summary(MODELOBJECT)$varcomp`
    + Residual Plot per $t_j$: `plot(dat3$tj,residuals(MODELOBJECT, type=("pearson")))`
    

## Standard Model
```{r Standard Model, results="hide", fig.show="hide"}
standard <- asreml(fixed   = adjmean ~ ri + tj,
                   random  = ~ G + L + Y + L:G + Y:G + G:Env,
                   ran.order = "user",
                   weights = w,
                   family  = asreml.gaussian(dispersion=1.0),
                   control = asreml.control(pworkspace=8e10,workspace=8e8,maxiter=100),
                   data    = dat3)
```

```{r Standard out, echo=FALSE}
paste("AIC= ", round(-2*standard$loglik + 2*length(standard$gammas),2),sep="")
wald(standard, denDF = c("none"),ssType = c("incremental","conditional"))[,c(1,3,4)]
summary(standard)$varcomp[,c(2,3,5)]
plot(dat3$tj,residuals(standard, type=("pearson")))
```

## Shukla's Stability Variance
```{r Shukla Model, results="hide", fig.show="hide"}
shukla <- asreml(fixed   = adjmean ~ ri + tj,
                    random  = ~ G + L + Y + L:G + Y:G + Y:L + at(G):Env, 
                    ran.order = "user",
                    weights = w,
                    family  = asreml.gaussian(dispersion=1.0),
                    control = asreml.control(pworkspace=8e10,workspace=8e8,maxiter=100),
                    data    = dat3)
```

```{r Shukla format, echo=FALSE}
shukla_vc  <- summary(shukla)$varcomp[2:3]
  vc_names <- rownames(shukla_vc)
  vc_names <- substr(vc_names,1,regexpr('!',vc_names)-1)
  vc_names <- gsub(")"," ",vc_names)
  vc_names <- gsub("at\\(G, ","",vc_names)
shukla_vc$VC <- as.factor(vc_names)
shukla_vc$yesno <- "no"
shukla_vc[which(regexpr(":Env",shukla_vc$VC)>0),"yesno"] <- "yes"
shukla_vc$StabVar <- NA
shukla_vc[which(shukla_vc$yesno=="yes"),"StabVar"] <- gsub(" :Env","",shukla_vc[which(shukla_vc$yesno=="yes"),"VC"])   
shukla_vc$StabVar   <- as.factor(shukla_vc$StabVar)
rownames(shukla_vc) <- NULL
shukla_vc <- shukla_vc[,c(3,1,2,5)]
colnames(shukla_vc) <- c("VC","Estimate","StdErr","Variety")
shukla_out <- subset(shukla_vc, is.na(shukla_vc$Variety)==F)
shukla_out <- shukla_out[,c("Variety","Estimate","StdErr")]
shukla_out[,c(2,3)] <- round(shukla_out[,c(2,3)],4)
```

```{r Shukla Out, echo=FALSE}
paste("AIC= ", round(-2*shukla$loglik + 2*length(shukla$gammas),2),sep="")
wald(shukla, denDF = c("none"),ssType = c("incremental","conditional"))[,c(1,3,4)]
shukla_out
plot(dat3$tj,residuals(shukla, type=("pearson")))
```

## Factor-Analytic Model
```{r FA Model, results="hide", fig.show="hide"}
FAmod <- asreml(fixed   = adjmean ~ ri + tj,
                random  = ~ G + L + Y + L:G + Y:G + fa(G):Env,
                ran.order = "user",
                weights = w,
                family  = asreml.gaussian(dispersion=1.0),
                control = asreml.control(pworkspace=8e10,workspace=8e8,maxiter=100),
                data    = dat3)
```

```{r FA format, echo=FALSE}
FA_vc <- summary(FAmod)$varcomp[2:3]
FA_vc$VC <- rownames(FA_vc)
rownames(FA_vc) <- NULL
FA_vc$isg   <- regexpr(":Env!G.",FA_vc$VC)
FA_vc$isfa  <- regexpr(".fa1",FA_vc$VC) 
FA_vc$group <- NA
FA_vc[which(FA_vc$isg>1) ,"group"] <- "sigma"
FA_vc[which(FA_vc$isfa>1),"group"] <- "lambda"
FA_vc[which(is.na(FA_vc$group)==T),"VC"] <- substr(FA_vc[which(is.na(FA_vc$group)==T),"VC"],1,regexpr('!',FA_vc[which(is.na(FA_vc$group)==T),"VC"])-1)
FA_vc$VC <- gsub("fa\\(G):Env!G.","",FA_vc$VC)
FA_vc$VC <- gsub(".var","",FA_vc$VC)
FA_vc$VC <- gsub(".fa1","",FA_vc$VC)
FA_vc<-FA_vc[,c(3,1,2,6)]
colnames(FA_vc) <- c("VC","Estimate","StdErr","group")

# Transposed Output
G_var  <- subset(FA_vc, group=="sigma")[,-c(4)]
colnames(G_var) <- c("VC","Var_Est","Var_SE")
G_lam  <- subset(FA_vc, group=="lambda")[,-c(4)]
colnames(G_lam) <- c("VC","Lambda_Est","Lambda_SE")
FA_out <- merge(G_var, G_lam, by="VC")
FA_out[,c(2,3,4,5)] <- round(FA_out[,c(2,3,4,5)],4)
```

```{r FA Out, echo=FALSE}
paste("AIC= ", round(-2*FAmod$loglik + 2*length(FAmod$gammas),2),sep="")
wald(FAmod, denDF = c("none"),ssType = c("incremental","conditional"))[,c(1,3,4)]
FA_out
plot(dat3$tj,residuals(FAmod, type=("pearson")))
```

## 5-Year-Chunk Variances
```{r fiveyr Model, results="hide", fig.show="hide"}
dat3$fiveyr <- as.factor(ceiling((dat3$tj-2000)/5)) #Create three 5-year groups from tj

fiveyr <- asreml(fixed   = adjmean ~ ri + tj,
                 random  = ~ G + L + Y + L:G + Y:G + diag(fiveyr):Env:G,
                 ran.order = "user",
                 weights = w,
                 family  = asreml.gaussian(dispersion=1.0),
                 control = asreml.control(pworkspace=8e10,workspace=8e8,maxiter=100),
                 data    = dat3)
```


```{r fiveyr Out, echo=FALSE}
paste("AIC= ", round(-2*fiveyr$loglik + 2*length(fiveyr$gammas),2),sep="")
wald(fiveyr, denDF = c("none"),ssType = c("incremental","conditional"))[,c(1,3,4)]
summary(fiveyr)$varcomp[,c(2,3,5)]
plot(dat3$tj,residuals(fiveyr, type=("pearson")))
```
